dataset header cleanup
