+++ src/components/DatasetNameHolder.css
@@ -1,0 +1,32 @@
+#dataset_name_holder{
+    /*position: absolute;
+    margin-top: 170px;
+    left: 0;
+    float: left;
+    margin-left: 25px;*/
+    position: relative;
+    margin-top: 0px;
+    left: 0;
+    top: 0;
+    float: left;
+    min-width: 100%;
+    margin-left: 0px;
+    padding-left: 0;
+    margin-right: 0;
+    padding-right: 0;
+}
+
+#text_title{
+    margin: 0;
+}
+
+#title_tag{
+    color: gray;
+}
+
+.subtitle_tag, .subtitle_content{
+    color: gray;
+    font-size:medium;
+    margin: 0;
+    padding: 0;
+}
